Move SourceLocation and its calculation from the deprecated `Fixie.TestAdapter` project to the `Fixie` project as a step towards dropping `Fixie.TestAdapter`.
