+++ src/Fixie.Tests/TestAdapter/SourceLocationProviderTests.cs
@@ -1,4 +1,4 @@
-﻿using Fixie.TestAdapter;
+﻿using Fixie.Internal;
 using static Fixie.Tests.Utility;
 
 namespace Fixie.Tests.TestAdapter;
